Add files via upload
Deploys user-created syslog configuration files to multiple IP addresses defined in a list the user creates. The script will connect to each device, check hostname and platform, and will do different functions depending on which conditions the user configures in the script. This script is designed to be used for teams that have multiple syslog servers spread out over their environment, but you can customize the script to match your needs.